allow more than 2^31 - 1 seeds to be checked

+++ src/main/java/com/mvc/allarthropods/Main.java
@@ -13,7 +13,7 @@ public class Main {
         final ChunkRand chunkRand = new ChunkRand();
         System.out.println("Starting seed finding...");
         FileWriter output = new FileWriter("./output.txt");
-        int seedsChecked = 0;
+        long seedsChecked = 0;
         int seedMatches = 0;
         long nextTime = 0;
         long currentTime;
